Check narrowing conversions in PSAP2 and ACSAP.
Copy presentation token ownership without going through int, keep
pb_flags in short, range-check context offsets and DASE string
lengths, and set ACB_RELEASE via unsigned 16-bit flag helpers so the
0x8000 bit is preserved in a signed short.

+++ psap2/psapinitiate.c
@@ -265,7 +265,7 @@ no_mem:
 	}
 
 	if (data && ndata > 0) {
-		int dflt = pb -> pb_flags;
+		short dflt = pb -> pb_flags;
 		pb -> pb_flags &= ~PB_DFLT; /* don't be clever for this pdu */
 		if ((normal -> user__data = info2ppdu (pb, pi, data, ndata,
 											   PPDU_CP)) == NULL)
